Organize init of api package
